+++ src/main/java/roomescape/auth/AuthenticationPrincipalArgumentResolver.java
@@ -8,7 +8,7 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 import roomescape.auth.dto.LoginMember;
-import roomescape.auth.service.AuthService;
+import roomescape.auth.application.AuthService;
 
 @Component
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
