+++ src/main/java/roomescape/auth/application/AuthService.java
@@ -1,4 +1,4 @@
-package roomescape.auth.service;
+package roomescape.auth.application;
 
 import org.springframework.stereotype.Service;
 import roomescape.auth.JwtTokenProvider;
